ui/requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
gevent==1.4.0
graphqlclient==0.2.4
greenlet==0.4.15
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pyjade==4.0.0
PyYAML==5.3
requests==2.22.0
six==1.14.0
urllib3==1.25.7
Werkzeug==0.16.0
